backend/app/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pydantic-settings
langchain
langchain-community
langchain-text-splitters
langchain-qdrant
langgraph
langchain-gigachat
fastembed
qdrant-client
pdfplumber
sqlalchemy 
psycopg2-binary 
python-jose[cryptography] 
passlib[bcrypt] 
pydantic[email]